// src/features/course-v1/model/types.ts
export type GradeType = 1 | 2 | 3;
export type EnvType =
  | 'PARK'
  | 'TRAIL'
  | 'TRACK'
  | 'URBAN'
  | 'BEACH'
  | 'MOUNTAIN'
  | 'FOREST'
  | 'RIVER';
// | 'ETC';
export type EnvTypeName =
  | '공원'
  | '산책로'
  | '트랙'
  | '도심'
  | '해변'
  | '산'
  | '숲'
  | '강';
// | '기타';
export type ShapeType = 'LOOP' | 'LINEAR' | 'OUT_AND_BACK' | 'ART'; // | 'ETC';
export type ShapeTypeName = '순환' | '직선' | '왕복' | '아트'; // | '기타';
// ...
export type CategoryCode = 'COURSE_GOOD' | 'ENVIRONMENT_GOOD' | 'FACILITY_GOOD';